CI: clear golangci lint findings
Four issues surfaced by golangci-lint v1.64.8 after the gofmt fix:

- internal/scan/scan_test.go (unconvert): int32(int32(...)) and
  int64(int64(...)) redundant casts on Avro map values.
- internal/catalog/hive/hive.go (unused): defaultClient struct field
  was never populated or read; removed.
- internal/writepattern/classifier.go (ineffassign): initial
  `pat := WritePatternUnknown` was always overwritten by the switch
  below (including a default arm). Switched to `var pat`.
- internal/catalog/hive/{hive,kerberos}.go (staticcheck SA1019):
  thrift.NewTSocket is deprecated. Switched to NewTSocketConf(addr, nil)
  which matches the current apache/thrift API surface.

Behaviour unchanged. Build + vet + go test all clean.

+++ internal/scan/scan_test.go
@@ -130,10 +130,10 @@ func buildTableFixture(t testing.TB, dataFiles, deleteFiles int, dataFileSize in
 		"sequence_number":      int64(1),
 		"min_sequence_number":  int64(1),
 		"added_snapshot_id":    int64(100),
-		"added_files_count":    int32(int32(dataFiles + deleteFiles)),
+		"added_files_count":    int32(dataFiles + deleteFiles),
 		"existing_files_count": int32(0),
 		"deleted_files_count":  int32(0),
-		"added_rows_count":     int64(int64(dataFiles * 100)),
+		"added_rows_count":     int64(dataFiles * 100),
 		"existing_rows_count":  int64(0),
 		"deleted_rows_count":   int64(0),
 	}})
